@@ -5,7 +5,7 @@
     "keywords": ["xbbcode","bbcode"],
     "license": "GPL-2.0-or-later",
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "geshi/geshi": "^1.0.9.1"
     },
     "require-dev": {
