@@ -32,10 +32,7 @@ class Td extends Tag
 {
     public const BEHAVIOUR = 'td';
 
-    /**
-     * @return Attributes
-     */
-    protected function getAttributes()
+    protected function getAttributes(): Attributes
     {
         $attr = parent::getAttributes();
 
